Revert "Refactor main"
