+++ collision/src/main/java/module-info.java
@@ -2,7 +2,4 @@ module Collision {
     exports collision;
     requires common;
     requires java.desktop;
-
-    uses interfaces.IProcessing;
 }
-
