+++ common/pom.xml
@@ -10,8 +10,6 @@
         <version>1.0-SNAPSHOT</version>
     </parent>
 
-    <packaging>jar</packaging>
-
     <groupId>peters.car</groupId>
     <artifactId>common</artifactId>
     <version>1.0-SNAPSHOT</version>
