+++ common/src/main/java/utilities/Types.java
@@ -8,4 +8,5 @@ public enum Types {
     BULLET,
     OBSTACLE,
     UNDEFINED;
+
 }
