+++ core/pom.xml
@@ -26,6 +26,12 @@
         </dependency>
         <dependency>
             <groupId>peters.car</groupId>
+            <artifactId>player</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
+        <dependency>
+            <groupId>peters.car</groupId>
             <artifactId>map</artifactId>
             <version>1.0-SNAPSHOT</version>
             <scope>compile</scope>
