+++ core/src/main/java/module-info.java
@@ -1,7 +1,7 @@
-import interfaces.IDrawable;
 import utilities.SPIlocator;
 
 module core {
+    requires player;
     requires gameEngine;
     requires common;
     requires map;
@@ -9,5 +9,4 @@ module core {
     uses interfaces.IPlugin;
     uses abstractClasses.Entity;
     uses SPIlocator;
-    uses IDrawable;
 }
