+++ gameEngine/pom.xml
@@ -25,7 +25,7 @@
             <version>1.0-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
-    </dependencies>
 
+    </dependencies>
 
 </project>
