+++ gameEngine/src/main/java/module-info.java
@@ -1,6 +1,5 @@
 module gameEngine {
     exports gameEngine;
-
     requires common;
     requires java.desktop;
 }
