+++ map/src/main/java/map/Map.java
@@ -4,9 +4,10 @@ import abstractClasses.Entity;
 import interfaces.IDrawable;
 import interfaces.IGameEngine;
 import interfaces.IPlugin;
-import utilities.GameData;
 import utilities.Shapes;
 import utilities.Types;
+
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.image.BufferedImage;
@@ -20,6 +21,7 @@ public class Map extends Entity implements IDrawable, IPlugin {
     private Tile grass;
     private Tile earth;
     private Tile obstacle;
+
     private ArrayList <Shapes> shapesArray;
 
     // combinedTiles:
@@ -75,25 +77,20 @@ public class Map extends Entity implements IDrawable, IPlugin {
     }
 
     @Override
+    public Entity create(IGameEngine gameEngine) {
+        return null;
+    }
+
+    @Override
+    public Entity delete(IGameEngine gameEngine) {
+        return null;
+    }
+
+
+    @Override
     public void draw(Graphics2D g, JPanel panel) {
 
         g.drawImage(getSprite().getImage(),0,0,panel);
 
-    }
-
-    @Override
-    public Entity create(GameData gameEngine) {
-        Entity newMap;
-        try {
-            newMap = new Map();
-        } catch (IOException e) {
-            throw new RuntimeException(e);
-        }
-        return newMap;
-    }
-
-    @Override
-    public Entity delete(GameData gameEngine) {
-        return null;
     }
 }
