+++ player/src/main/java/module-info.java
@@ -5,6 +5,4 @@ module player {
 
     provides interfaces.IPlugin with player.PlayerPlugin;
     provides abstractClasses.Entity with player.PlayerPlugin;
-    provides interfaces.IDrawable with player.PlayerPlugin;
-    provides interfaces.IProcessing with player.PlayerPlugin;
 }
